+++ pyslvs_ui/__main__.py
@@ -3,7 +3,7 @@
 """Launch script from module level."""
 
 __author__ = "Yuan Chang"
-__copyright__ = "Copyright (C) 2016-2019"
+__copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
@@ -54,6 +54,7 @@ def main() -> None:
     del sp
     logger.info(f"Startup with: {process_time() - t0:.02f}s")
     qt_exit_code = _app.exec_()
+    del lf
     shutdown()
     exit(qt_exit_code)