+++ pyslvs_ui/widgets/__init__.py
@@ -5,12 +5,13 @@ that design without Qt designer.
 """
 
 __author__ = "Yuan Chang"
-__copyright__ = "Copyright (C) 2016-2019"
+__copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
 from .main_base import MainWindowBase, Preferences
 from .tables import PointArgs, LinkArgs, PointTableWidget, LinkTableWidget
+from .rotatable import QRotatableView
 from .undo_redo import (
     AddTable,
     AddPath,
@@ -34,6 +35,7 @@ __all__ = [
     'LinkArgs',
     'PointTableWidget',
     'LinkTableWidget',
+    'QRotatableView',
     'AddTable',
     'AddPath',
     'AddStorage',