+++ pyslvs_ui/widgets/canvas_base.py
@@ -5,7 +5,7 @@
 from __future__ import annotations
 
 __author__ = "Yuan Chang"
-__copyright__ = "Copyright (C) 2016-2019"
+__copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
@@ -120,7 +120,7 @@ class MainCanvasBase(BaseCanvas, ABC):
     selected_tips = Signal(QPoint, str)
     selected_tips_hide = Signal()
     free_moved = Signal(tuple)
-    noselected = Signal()
+    no_selected = Signal()
     alt_add = Signal(float, float)
     doubleclick_edit = Signal(int)
     zoom_changed = Signal(int)
@@ -616,7 +616,7 @@ class MainCanvasBase(BaseCanvas, ABC):
                     and km != Qt.ControlModifier
                     and km != Qt.ShiftModifier
                 ):
-                    self.noselected.emit()
+                    self.no_selected.emit()
             else:
                 if km == Qt.AltModifier:
                     # Add Point
@@ -629,7 +629,7 @@ class MainCanvasBase(BaseCanvas, ABC):
                     and km != Qt.ControlModifier
                     and km != Qt.ShiftModifier
                 ):
-                    self.noselected.emit()
+                    self.no_selected.emit()
         self.selected_tips_hide.emit()
         self.selector.release()
         self.update()
@@ -660,7 +660,7 @@ class MainCanvasBase(BaseCanvas, ABC):
                     if selection:
                         self.selected.emit(selection, False)
                     else:
-                        self.noselected.emit()
+                        self.no_selected.emit()
 
                     self.selected_tips.emit(
                         event.globalPos(),